update - 08/12
more functionality is now broken up, and each of the python files now performs its intended function more purely. a lot of the browser requests have instead been shifted over to javascript/ajax. unfortunately, flask does not have any way to send post requests, but only how to interpret them. therefore most of the data loading is now initiated by app.js, which calls on views, which calls on dbAccess, to finish the function.

most of the data passed around is now passed through the browser via a post request, and most of the data passed in between is now in json format, for the sake of compressing data