update - 08/14
modified html so that and/or selector is added when additional attributes are added. the visual tool is now fully functioning, and can complete queries within the scope of the UI - meaning that the visual tool is capable of selecting specific columns from any table in a database while applying constraints to any of the columns. also contains more extensive commenting throughout