update - 08/18
added comments to many of the files to explain functions

added db change page so users can change which database they are looking at by logging back in with a server login

added buttons to top to reset query tool and reset which database you're accessing. again, changing databases requires a login to the server, so keep that in mind when setting your default db.

changed background of page to light blue :-)