update - 08/19 (the security update!)
updated dom so that 'reset query' button is only shown on query page. it doesn't really change much, as doing so before would just leave you on the default database, but i suppose it better locks down the application

moved default connection (for fetching list of dbs at home) into the loadDBList() method. trying to minimize active connections, as the user base in mind uses a lot of mobile computers (microsoft surface), so we don't want those to leave the server hanging

added "injectionAlert()" to detect key words in a query, which are operations that should not be executed by this tool at all. this is a read only tool, so there's no need to execute anything that modifies the data in the database.